backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
groq==0.15.0
pydantic==2.10.4
python-dotenv==1.0.1
numpy<2.0
sentence-transformers==3.3.1
PyMuPDF==1.25.3
